Added vCard download.
