+++ egov/contactdirectory/profiles/default/actions.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0"?>
+<object name="portal_actions" meta_type="Plone Actions Tool"
+    xmlns:i18n="http://xml.zope.org/namespaces/i18n">
+
+    <object name="document_actions" meta_type="CMF Action Category">
+        <object name="vcard_download" meta_type="CMF Action" i18n:domain="egov.contactdirectory">
+            <property name="title" i18n:translate="">Download vCard</property>
+            <property name="description" i18n:translate="">Download a Contact as a vCard.</property>
+            <property name="url_expr">string:${object_url}/download</property>
+            <property name="link_target"></property>
+            <property name="icon_expr"></property>
+            <property name="available_expr">python:object.portal_type == 'Contact'</property>
+            <property name="permissions">
+                <element value="View" />
+            </property>
+            <property name="visible">True</property>
+        </object>
+    </object>
+
+</object>
