+++ egov/contactdirectory/profiles/default/metadata.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0"?>
 <metadata>
-  <version>1500</version>
   <dependencies>
       <dependency>profile-ftw.geo:default</dependency>
   </dependencies>
