+++ egov/contactdirectory/upgrades/20150805143611_add_v_card_export_actions/upgrade.py
@@ -1,0 +1,9 @@
+from ftw.upgrade import UpgradeStep
+
+
+class AddVCardExportActions(UpgradeStep):
+    """Add v card export actions.
+    """
+
+    def __call__(self):
+        self.install_upgrade_profile()
